Create app-core package and update dashboard to use shared components
