+++ apps/dashboard/tailwind.config.ts
@@ -3,7 +3,7 @@ module.exports = {
   content: [
     "./src/app/**/*.{js,ts,jsx,tsx}",
     "./src/components/**/*.{js,ts,jsx,tsx}",
-    "'../../packages/appCore/src/**/*.{js,ts,jsx,tsx}'",
+    "../../packages/app-core/src/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {},
