+++ packages/app-core/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../../apps/dashboard/tsconfig.json",
+  "compilerOptions": {
+    "composite": true,
+    "rootDir": "./src",
+    "outDir": "./dist",
+    "baseUrl": "."
+  },
+  "include": ["src/**/*.ts", "src/**/*.tsx"],
+  "exclude": ["node_modules", "dist"]
+}
+
